Unit tests successfully passed either for Python 2 or Python 3
