+++ build_bot/build_bot.py
@@ -1,5 +1,5 @@
-from common.cmd_utils import shell
-from ibuild_bot import IBuildBot
+from build_bot.common.cmd_utils import shell
+from build_bot.ibuild_bot import IBuildBot
 
 
 class BuildBot(IBuildBot):
