+++ build_bot/unit_tests/conftest.py
@@ -1,4 +1,7 @@
 import sys
+from build_bot.languages.cpp_language import CPPLanguage
+from build_bot.languages.cs_language import CSLanguage
+from build_bot.languages.fp_language import FPLanguage
 
 if sys.version_info > (3, 0):
     # Python 3 code in this block
@@ -10,10 +13,6 @@ else:
 import os
 import shutil
 import pytest
-from cpp_language import CPPLanguage
-from cs_language import CSLanguage
-from fp_language import FPLanguage
-
 
 @pytest.fixture(scope="class")
 def test_build_directory_fixture(request):
