+++ build_bot/unit_tests/test_cpp_build_bot.py
@@ -1,9 +1,10 @@
 import os
 import unittest
 import pytest
-from build_bot import BuildBot
-from cmd_utils import shell, split_lines
-from cpp_language import CPPLanguage
+from build_bot.build_bot import BuildBot
+from build_bot.common.cmd_utils import shell
+from build_bot.common.cmd_utils import split_lines
+from build_bot.languages.cpp_language import CPPLanguage
 
 test_name = __name__
