+++ build_bot/unit_tests/test_cs_language.py
@@ -1,6 +1,6 @@
 import os
 import unittest
-from cs_language import CSLanguage
+from build_bot.languages.cs_language import CSLanguage
 
 
 class CSLanguageTests(unittest.TestCase):
